RAG Chatbot/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
cohere
qdrant-client
psycopg2-binary